+++ finishScreen.c
@@ -9,8 +9,8 @@ int backButtonSwitch = 0;
 void DrawFinishScreen() {
     BeginDrawing();
 
-    DrawTexturePro(currentBeatmap->background, (Rectangle) { 0, 0, screenWidth, screenHeight },
-                   (Rectangle) { 0, 0, screenWidth, screenHeight }, (Vector2) { 0, 0 }, 0,
+    DrawTexturePro(currentBeatmap->background, (Rectangle) { 0, 0, (float)screenWidth, (float)screenHeight },
+                   (Rectangle) { 0, 0, (float)screenWidth, (float)screenHeight }, (Vector2) { 0, 0 }, 0,
                    WHITE);
     DrawRectangleGradientV(0, 0, screenWidth, screenHeight, BLACK, BLANK);
 
@@ -20,7 +20,7 @@ void DrawFinishScreen() {
     DrawText(TextFormat("Good: %i", goodCounter), 100, 120, 20, WHITE);
     DrawText(TextFormat("Misses: %i", missCounter), 100, 140, 20, WHITE);
 
-    backButtonSwitch = GuiButton((Rectangle){ 100, screenHeight - 200, screenWidth / 3, screenHeight / 7 },
+    backButtonSwitch = GuiButton((Rectangle){ 100, (float)screenHeight - 200, (float)screenWidth / 3, (float)screenHeight / 7 },
                                            "Back");
 
     EndDrawing();
